% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_algorithm_helpers.R
\name{aggregate_ga_results}
\alias{aggregate_ga_results}
\title{Aggregate GA Results from Multiple JSON Files returned by run_ga_batch}
\usage{
aggregate_ga_results(folder_path = "results")
}
\arguments{
\item{folder_path}{Path to the folder containing result JSON files.}
}
\value{
A data frame of unique cocktails with their scores.
}
\description{
Aggregate GA Results from Multiple JSON Files returned by run_ga_batch
}
