% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_algorithm_helpers.R
\name{filter_out_cocktails}
\alias{filter_out_cocktails}
\title{Filter a list of cocktails accoring to the mean depth of nodes inside
each cocktail. A cocktail is kept if the mean depth of his nodes is greater
than \code{mean_depth_cocktail} (default = 3). This function aim is to provide
more interpretable cocktail to the user.}
\usage{
filter_out_cocktails(
  results,
  tree_depth,
  mean_depth_cocktail = 3,
  one_index = FALSE
)
}
\arguments{
\item{results}{results returned by function \code{summarize_ga_pipeline}. Or a
list of string of 0 indexed vector of tree nodes. (If nodes are 1-indexed, make sure that the parameter
one_index is set to TRUE).}

\item{tree_depth}{The tree structure depth.}

\item{mean_depth_cocktail}{The desired minimal mean depth of nodes in a cocktail.}

\item{one_index}{A boolean telling if the cocktails are 1-indexed or 0-indexed. Default is FALSE}
}
\value{
The filtered list of cocktail
}
\description{
Filter a list of cocktails accoring to the mean depth of nodes inside
each cocktail. A cocktail is kept if the mean depth of his nodes is greater
than \code{mean_depth_cocktail} (default = 3). This function aim is to provide
more interpretable cocktail to the user.
}
