% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pwp_combination_flag_cpp}
\alias{pwp_combination_flag_cpp}
\title{Construct one hierarchical combination indicator}
\usage{
pwp_combination_flag_cpp(
  combination,
  index_base,
  patient_data,
  node_column,
  target_column,
  id_column,
  tree,
  depth_column,
  upper_bound_column,
  name_column
)
}
\description{
Construct one hierarchical combination indicator
}
\keyword{internal}
