% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mcmc_pwp_cpp}
\alias{run_mcmc_pwp_cpp}
\title{Internal C++ engine for PWP Rao MCMC}
\usage{
run_mcmc_pwp_cpp(
  patient_data,
  node_column,
  target_column,
  id_column,
  tree,
  depth_column,
  upper_bound_column,
  name_column,
  score_context,
  epochs = 1e+05,
  temperature = 1,
  n_results = 20,
  cocktail_size = 2,
  prob_type1 = 0.01,
  beta = 20,
  max_score = 50,
  seed = 1L,
  verbose = FALSE,
  burn_in = 0,
  store_trace = FALSE
)
}
\description{
Use \code{run_pwp_mcmc()} from R.
}
\keyword{internal}
