% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score_pwp_combinations_cpp}
\alias{score_pwp_combinations_cpp}
\title{Evaluate PWP Rao statistics for specified combinations}
\usage{
score_pwp_combinations_cpp(
  combinations,
  index_base,
  patient_data,
  node_column,
  target_column,
  id_column,
  tree,
  depth_column,
  upper_bound_column,
  name_column,
  score_context
)
}
\description{
Evaluate PWP Rao statistics for specified combinations
}
\keyword{internal}
