% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treehunt-package.R
\docType{package}
\name{treehunt-package}
\alias{treehunt}
\alias{treehunt-package}
\title{Hierarchical combination search and score reference distributions}
\description{
Provides Markov chain Monte Carlo (MCMC) sampling and genetic algorithms for searching combinations of nodes in hierarchical trees. Includes scores for binary outcomes, continuous outcomes, and recurrent events, with applications to pharmacovigilance and hospital data. Importance weighting estimates uniform combination score references from tempered MCMC samples and reports weight-based effective sample sizes. Computational routines are implemented in C++ via 'Rcpp'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JulesBa-Git/treehunt}
  \item Report bugs at \url{https://github.com/JulesBa-Git/treehunt/issues}
}

}
\author{
\strong{Maintainer}: Jules Bangard \email{jules.bangard@etu.unistra.fr}

Authors:
\itemize{
  \item Jules Bangard \email{jules.bangard@etu.unistra.fr}
}

}
\keyword{internal}
