% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwp_rao.R
\name{decode_treehunt_solutions}
\alias{decode_treehunt_solutions}
\title{Decode treehunt node indexes}
\usage{
decode_treehunt_solutions(
  combinations,
  tree,
  index_base = 0L,
  code_column = "Code"
)
}
\arguments{
\item{combinations}{List of node-index vectors.}

\item{tree}{Tree data frame.}

\item{index_base}{Index base of `combinations`.}

\item{code_column}{Column containing node labels.}
}
\description{
Decode treehunt node indexes
}
