% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_dissimilarity_of_list}
\alias{get_dissimilarity_of_list}
\title{Compute the dissimilarity matrix on a list of cocktails}
\usage{
get_dissimilarity_of_list(
  cocktail_list,
  patient_data,
  node_column,
  target_column,
  tree,
  depth_column,
  upper_bound_column = NULL,
  name_column = NULL,
  score_type = "wilcoxon"
)
}
\description{
Compute the dissimilarity matrix on a list of cocktails
}
