% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_taker}
\alias{get_taker}
\title{Temporary helpers : Used to get the takers of a cocktail or cocktail list in
order to get a sense of the used medications as the ndc are mapped to ATC4}
\usage{
get_taker(
  cocktail_list,
  patient_data,
  node_column,
  target_column,
  id_column,
  hadm_column,
  tree,
  depth_column,
  upper_bound_column = NULL,
  name_column = NULL
)
}
\description{
Temporary helpers : Used to get the takers of a cocktail or cocktail list in
order to get a sense of the used medications as the ndc are mapped to ATC4
}
