% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_algorithm_helpers.R
\name{map_cocktail_names}
\alias{map_cocktail_names}
\title{Map Cocktail Indices to Names
TODO : Results of GA are 0 based index for now, they will be change in a near future
when this is done, remove the cpp to R index conversion}
\usage{
map_cocktail_names(aggregated_df, new_tree)
}
\arguments{
\item{aggregated_df}{The data frame returned by aggregate_ga_results.}

\item{new_tree}{The data frame containing the tree structure and $Name column.}
}
\value{
The same data frame with a new 'cocktail_names' column.
}
\description{
Map Cocktail Indices to Names
TODO : Results of GA are 0 based index for now, they will be change in a near future
when this is done, remove the cpp to R index conversion
}
