% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_algorithm_helpers.R
\name{run_ga_batch}
\alias{run_ga_batch}
\title{Run Multiple Genetic Algorithm Replicates from Configuration}
\usage{
run_ga_batch(
  config_path,
  patient_data,
  tree_df,
  seed_population = NULL,
  replicates = 5,
  output_dir = "results"
)
}
\arguments{
\item{config_path}{Path to a JSON file containing GA hyperparameters.}

\item{patient_data}{The data frame containing patient/node observations.}

\item{tree_df}{The data frame defining the tree structure.}

\item{replicates}{Number of times to run each configuration. Default is 5.}

\item{output_dir}{Directory where result JSONs will be saved.}
}
\description{
Run Multiple Genetic Algorithm Replicates from Configuration
}
